Create Clipboard.kt for clipboard-fix in JS

+++ core-library/src/main/kotlin/com/zimolab/monacofx/monaco/editor/event/mouse/interfaces/IMouseEvent.kt
@@ -19,7 +19,7 @@ interface IMouseEvent {
     val shiftKey: Boolean
     val altKey: Boolean
     val metaKey: Boolean
-    val timestamp: Int
+    val timestamp: Double
     @JsFunction
     fun preventDefault()
     @JsFunction
